Fixed implementation of Menu class after removing old Menu class and renaming BetterMenu to Menu.
