+++ src/main/java/com/infoshareacademy/App.java
@@ -1,6 +1,8 @@
 package com.infoshareacademy;
 
 import com.infoshareacademy.configuration.PropertiesLoader;
+import com.infoshareacademy.presentationlayer.Menu;
+import com.infoshareacademy.presentationlayer.ValuesScanner;
 
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicReference;
@@ -13,26 +15,26 @@ public class App
         PropertiesLoader propertiesLoader = new PropertiesLoader();
         propertiesLoader.loadProperties();
 
- /*       BetterMenu menu = new BetterMenu();
+        Menu menu = new Menu();
 
-        menu.addMenuOption(new BetterMenu.MenuOption().
+        menu.addMenuOption(new Menu.MenuOption().
                 setDescription("Properties Loader").
                 setMethod(()->{
                     PropertiesLoader p = new PropertiesLoader();
                     try {
-                        p.propertiesLoader();
+                        p.loadProperties();
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
                 }));
 
-        menu.addMenuOption(new BetterMenu.MenuOption().
+        menu.addMenuOption(new Menu.MenuOption().
                 setDescription("New Menu Option").
                 setMethod(()->{
                     System.out.println("This is a test. You've chosen '1'.");
         }));
 
-        menu.addMenuOption(new BetterMenu.MenuOption().
+        menu.addMenuOption(new Menu.MenuOption().
                 setDescription("Exit").
                 setMethod(()->{
                     stayInLoop.set(false);
@@ -41,15 +43,15 @@ public class App
         menu.displayMenu();
         while(stayInLoop.get()){
             menu.executeSelectedOption(ValuesScanner.scanIntegerInRange("Select the desired option", 0 , menu.getMenuSize()));
-        }*/
+        }
 
 
 
 
-/*        System.out.println( "Rekiny Finansjery" );
+        System.out.println( "Rekiny Finansjery" );
 
         NBPApiManager nbpApiManager = new NBPApiManager();
 
-        UsageExamplesCode.showExamples(nbpApiManager);*/
+        UsageExamplesCode.showExamples(nbpApiManager);
     }
 }
