+++ src/main/java/com/infoshareacademy/UsageExamplesCode.java
@@ -1,10 +1,6 @@
 package com.infoshareacademy;
 
-import com.infoshareacademy.presentationlayer.CollectionView;
-import com.infoshareacademy.data.ExchangeRatesArchiveTable;
 import com.infoshareacademy.presentationlayer.Menu;
-
-import java.time.LocalDate;
 
 public class UsageExamplesCode {
      interface Test{
@@ -18,7 +14,7 @@ public class UsageExamplesCode {
     public static void showExamples(NBPApiManager nbpApiManager){
 
         Menu menu = new Menu();
-        menu.showMenu();
+        menu.displayMenu();
          Runnable runnable = new Runnable() {
              @Override
              public void run() {
